requirements.txt
appnope==0.1.0
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
Django==3.0.2
django-debug-toolbar==2.2
django-extensions==2.2.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
Faker==4.0.1
idna==2.8
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
more-itertools==8.0.2
packaging==20.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
py==1.8.1
Pygments==2.5.2
pyparsing==2.4.6
pytest==5.3.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
transliterate==1.10.2
urllib3==1.25.7
wcwidth==0.1.8
